MNT: reduce number of implicit imports from toplevel __init__.py
Either hide or defer imports from the toplevel import
