+++ lib/matplotlib/pyplot.py
@@ -74,7 +74,7 @@ from matplotlib.scale import get_scale_names
 from matplotlib import cm
 from matplotlib.cm import _colormaps as colormaps
 from matplotlib.cm import register_cmap  # type: ignore
-from matplotlib.colors import _color_sequences as color_sequences
+from matplotlib._colors import _color_sequences as color_sequences
 
 import numpy as np
