+++ lib/matplotlib/rcsetup.py
@@ -24,7 +24,7 @@ import numpy as np
 
 from matplotlib import _api, cbook
 from matplotlib.cbook import ls_mapper
-from matplotlib.colors import Colormap, is_color_like
+from matplotlib._colors import Colormap, is_color_like
 from matplotlib._fontconfig_pattern import parse_fontconfig_pattern
 from matplotlib._enums import JoinStyle, CapStyle
